api-consumo/pyproject.toml
[tool.poetry]
name = "api-consumo"
version = "0.1.0"
description = ""
authors = ["Daniel Alexandre Gouveia <daniel.gouveia@serpro.gov.br>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"
pymongo = "^4.1.1"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
